m_requirements.txt
langchain
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
xmltodict    # for medical
requests
llama-index
typing
pydantic
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain-community
langchain-groq
sqlalchemy
langchain-core>=0.3.10,<0.4.0
langchain-community==0.3.2
pymysql  # Add to requirements.txt
requests